+++ src/main/java/faang/school/achievement/repository/AchievementRepository.java
@@ -6,6 +6,4 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface AchievementRepository extends JpaRepository<Achievement, Long> {
-
-    boolean existsByUserIdAndAchievementId(Long userId, Long achievementId);
 }
